Work around an issue with :drop
Running `:drop x` in one of several `x` windows can move you into
another one of them, so avoid calling it after opening a single file as
it's not actually necessary.  There is still an issue if opening
multiple files and the first one is already opened in another window.

Thanks to Hamza Tokuchi (a.k.a. Meuko).

Closes vifm/vifm.vim#71 on GitHub.